Split out JSON handling stuff from RegsGen2.
We need to isolate JSON handling code from RegsGen2 to be able to use it
from ds2.
