+++ Tools/JSObjects/Headers/JSObjects/JSObjects.h
@@ -11,8 +11,6 @@
 #ifndef __JSObjects_h
 #define __JSObjects_h
 
-#include "jsoncb.h"
-
 #include <cstdint>
 #include <cstdio>
 #include <map>
