+++ Tools/JSObjects/Sources/libjson/LICENSE_1_0.txt
