+++ Tools/JSObjects/Sources/libjson/qstring.c
