+++ Tools/JSObjects/Sources/libjson/qstring.h
