+++ Tools/RegsGen2/RegisterTemplate.h
@@ -11,7 +11,7 @@
 #ifndef __regsgen_RegisterTemplate_h
 #define __regsgen_RegisterTemplate_h
 
-#include "JSObjects.h"
+#include "JSObjects/JSObjects.h"
 #include "Definitions.h"
 
 class RegisterTemplate {
